+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Arbitrary.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
@@ -69,7 +70,9 @@ import Test.Cardano.Data (genNonEmptyMap)
 import Test.Cardano.Data.Arbitrary ()
 import Test.Cardano.Ledger.Alonzo.Arbitrary (genValidAndUnknownCostModels, genValidCostModel)
 import Test.Cardano.Ledger.Babbage.Arbitrary ()
+import Test.Cardano.Ledger.Binary.Random (QC (..))
 import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Core.Arbitrary (uniformSubMap)
 
 instance
   (Era era, Arbitrary (PParamsUpdate era)) =>
@@ -248,9 +251,12 @@ instance
 
 data ProposalsForEnactment era
   = ProposalsForEnactment
-      (Proposals era)
-      (Seq.Seq (GovActionState era))
-      (Set.Set (GovActionId (EraCrypto era)))
+  { pfeProposals :: Proposals era
+  , pfeToEnact :: Seq.Seq (GovActionState era)
+  , pfeToRemove :: Set.Set (GovActionId (EraCrypto era))
+  , pfeToRetain :: Set.Set (GovActionId (EraCrypto era))
+  -- ^ Those that are retained can only be the ones that don't have any lineage
+  }
   deriving (Show, Eq)
 
 instance
@@ -259,22 +265,41 @@ instance
   where
   arbitrary = do
     ps <- genProposals @era (2, 50)
+    let gasHasNoLineage gas =
+          case gasAction gas of
+            InfoAction {} -> True
+            TreasuryWithdrawals {} -> True
+            _ -> False
+        hasNoLineage gaId =
+          case proposalsLookupId gaId ps of
+            Nothing -> error $ "Expected " ++ show gaId ++ " in generated proposals"
+            Just gas -> gasHasNoLineage gas
     pparamUpdates <- chooseLineage grPParamUpdateL ps Seq.Empty
     hardForks <- chooseLineage grHardForkL ps Seq.Empty
     committees <- chooseLineage grCommitteeL ps Seq.Empty
     constitutions <- chooseLineage grConstitutionL ps Seq.Empty
+    noLineageMap <- uniformSubMap Nothing (Map.filter gasHasNoLineage $ proposalsActionsMap ps) QC
+    noLineage <- Seq.fromList <$> shuffle (Map.elems noLineageMap)
     sequencedGass <-
       sequenceLineages
         ( Seq.filter
             (not . Seq.null)
-            (Seq.fromList [pparamUpdates, hardForks, committees, constitutions])
+            (Seq.fromList [pparamUpdates, hardForks, committees, constitutions, noLineage])
         )
         Seq.Empty
-    let expiredGais =
+    let notEnacted =
           Set.fromList (toList $ proposalsIds ps)
             `Set.difference` Set.fromList (gasId <$> toList sequencedGass)
-    pure $ ProposalsForEnactment ps sequencedGass expiredGais
+    let (retained, removedDueToConflict) = Set.partition hasNoLineage notEnacted
+    pure $
+      ProposalsForEnactment
+        { pfeProposals = ps
+        , pfeToEnact = sequencedGass
+        , pfeToRemove = removedDueToConflict
+        , pfeToRetain = retained
+        }
     where
+      -- Starting from the root select a path in the tree until a leaf is reached.
       chooseLineage ::
         (forall f. Lens' (GovRelation f era) (f (GovPurposeId p era))) ->
         Proposals era ->
@@ -297,6 +322,7 @@ instance
                   chooseLineage govRelL ps (lineage Seq.:|> (proposalsActionsMap ps Map.! unGovPurposeId child))
       consumeHeadAtIndex :: Int -> Seq.Seq (Seq.Seq a) -> (a, Seq.Seq (Seq.Seq a))
       consumeHeadAtIndex idx ss = (ss `Seq.index` idx `Seq.index` 0, Seq.adjust' (Seq.drop 1) idx ss)
+      -- Mix lineages at random, while preserving relative order of each lineage
       sequenceLineages :: Seq.Seq (Seq.Seq a) -> Seq.Seq a -> Gen (Seq.Seq a)
       sequenceLineages lineages sequenced = case lineages of
         Seq.Empty -> pure sequenced
@@ -338,7 +364,7 @@ genProposals range = do
   go (def & pRootsL .~ fromPrevGovActionIds pgais) i
   where
     go :: Proposals era -> Int -> Gen (Proposals era)
-    go ps n
+    go !ps n
       | n <= 0 = pure ps
       | otherwise = do
           gas <- genGovActionState @era =<< genGovAction ps